refactor: add styles for rest page
