+++ src/components/rest-client/BodyEditor.tsx
@@ -1,9 +1,9 @@
 'use client';
 
-import { Form } from 'antd';
+import { Flex, Form } from 'antd';
 import { FormInstance, useWatch } from 'antd/es/form/Form';
 
-import CodeSpace from '@/components/CodeSpace';
+import CodeSpace from '@/components/rest-client/CodeSpace';
 import { FormValues } from '@/types/rest-client';
 
 type Props = {
@@ -15,14 +15,16 @@ export default function BodyEditor({ form }: Readonly<Props>) {
 
   return (
     <Form.Item name="body">
-      <CodeSpace
-        value={bodyValue || ''}
-        onChange={(value) => {
-          form.setFieldsValue({ body: value });
-        }}
-        height="200px"
-        language="json"
-      />
+      <Flex align="center" justify="center">
+        <CodeSpace
+          value={bodyValue || ''}
+          onChange={(value) => {
+            form.setFieldsValue({ body: value });
+          }}
+          height="200px"
+          language="json"
+        />
+      </Flex>
     </Form.Item>
   );
 }
